feat: move some functions into context
