+++ src/App.js
@@ -1,30 +1,14 @@
-import { useState } from "react";
-import { v4 as uuidv4 } from "uuid";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
 import Header from "./components/Header";
 import FeedbackList from "./components/FeedbackList";
 import FeedbackStats from "./components/FeedbackStats";
 import FeedbackForm from "./components/FeedbackForm";
-import FeedbackData from "./data/FeedbackData";
 import AboutPage from "./pages/AboutPage";
 import AboutIconLink from "./components/AboutIconLink";
 import { FeedbackProvider } from "./context/FeedbackContext";
 
 function App() {
-  const [feedback, setFeedback] = useState(FeedbackData);
-
-  const handleDelete = (id) => {
-    if (window.confirm("Are you sure delete this item?")) {
-      setFeedback(feedback.filter((el) => el.id !== id));
-    }
-  };
-
-  const handleAddFeedback = (newFeedback) => {
-    newFeedback.id = uuidv4();
-    setFeedback([newFeedback, ...feedback]);
-  };
-
   return (
     <FeedbackProvider>
       <Router>
@@ -36,9 +20,9 @@ function App() {
               path="/"
               element={
                 <>
-                  <FeedbackForm handleAdd={handleAddFeedback} />
+                  <FeedbackForm />
                   <FeedbackStats />
-                  <FeedbackList handleDelete={(id) => handleDelete(id)} />
+                  <FeedbackList />
                 </>
               }
             />
