+++ src/components/FeedbackForm.jsx
@@ -1,15 +1,17 @@
-import React, { useState } from "react";
+import React, { useState, useContext } from "react";
 
 import Card from "./shared/Card";
 import Button from "./shared/Button";
 import RatingSelect from "./RatingSelect";
+import FeedbackContext from "../context/FeedbackContext";
 
-function FeedbackForm({ handleAdd }) {
+function FeedbackForm() {
   const messageText = "Test must be least 10 characters!";
   const [text, setText] = useState("");
   const [btnDisabled, setBtnDisabled] = useState(true);
   const [message, setMessage] = useState("");
   const [rating, setRating] = useState(10);
+  const { handleAddFeedback } = useContext(FeedbackContext);
 
   const onTextChange = (event) => {
     if (text === "") {
@@ -39,7 +41,7 @@ function FeedbackForm({ handleAdd }) {
         rating,
       };
 
-      handleAdd(newFeedBack);
+      handleAddFeedback(newFeedBack);
       setText("");
     }
   };
