+++ src/components/FeedbackItem.jsx
@@ -1,8 +1,12 @@
 import PropTypes from "prop-types";
+import { useContext } from "react";
 import { FaTimes } from "react-icons/fa";
 import Card from "./shared/Card";
+import FeedbackContext from "../context/FeedbackContext";
 
-function FeedbackItem({ item, handleDelete }) {
+function FeedbackItem({ item }) {
+  const { handleDelete } = useContext(FeedbackContext);
+
   return (
     <Card>
       <div className="num-display">{item.rating}</div>
@@ -16,7 +20,6 @@ function FeedbackItem({ item, handleDelete }) {
 
 FeedbackItem.propTypes = {
   item: PropTypes.object.isRequired,
-  handleDelete: PropTypes.func.isRequired,
 };
 
 export default FeedbackItem;
